pyproject.toml
[build-system]
requires = ["setuptools>=65.0"]
build-backend = "setuptools.build_meta"

[project]
name = "dpsm-core"
version = "0.1.0"
description = "Common Core Foundation for DPSM experiments"
requires-python = ">=3.9"
license = {text = "Apache-2.0"}
authors = [
    {name = "OpenAI Research Automation", email = "noreply@example.com"}
]

dependencies = [
    "torch",
    "tqdm",
    "pyyaml",
    "matplotlib",
    "seaborn",
    "pandas",
    "transformers",
    "datasets",
    "sentencepiece",
    "rouge-score",
]
